Fix errors when building for a 32-bit x86 target.
Use size_t directly where possible, or cast to ptrdiff_t when required
as an index into an iterator.

Decorate methods bound to JavaScript with JSC_HOST_CALL to match the
required "fastcall" method signature.

+++ cobalt/browser/application.cc
@@ -770,7 +770,8 @@ void Application::UpdatePeriodicStats() {
   }
 #elif defined(OS_STARBOARD)
   int64_t used_cpu_memory = SbSystemGetUsedCPUMemory();
-  available_memory_ = SbSystemGetTotalCPUMemory() - used_cpu_memory;
+  available_memory_ =
+      static_cast<ssize_t>(SbSystemGetTotalCPUMemory() - used_cpu_memory);
   c_val_stats_.free_cpu_memory = available_memory_;
   c_val_stats_.used_cpu_memory = used_cpu_memory;
